refactor: move xblock handlers into base view
