+++ lms/djangoapps/ora_staff_grader/views.py
@@ -17,8 +17,7 @@ from lms.djangoapps.ora_staff_grader.errors import ERR_BAD_ORA_LOCATION, ERR_LOC
 from lms.djangoapps.ora_staff_grader.serializers import (
     InitializeSerializer, LockStatusSerializer, StaffAssessSerializer, SubmissionFetchSerializer, SubmissionStatusFetchSerializer
 )
-from lms.djangoapps.ora_staff_grader.utils import require_params
-from lms.djangoapps.ora_staff_grader.xblock_handlers_mixin import XblockHandlersMixin
+from lms.djangoapps.ora_staff_grader.utils import call_xblock_json_handler, require_params
 from openedx.core.djangoapps.content.course_overviews.api import get_course_overview_or_none
 from openedx.core.lib.api.authentication import BearerAuthenticationAllowInactiveUser
 
@@ -26,8 +25,8 @@ PARAM_ORA_LOCATION = 'oraLocation'
 PARAM_SUBMISSION_ID = 'submissionUUID'
 
 
-class StaffGraderBaseView(XblockHandlersMixin, RetrieveAPIView):
-    """ Base view for common setup between ESG views """
+class StaffGraderBaseView(RetrieveAPIView):
+    """ Base view for common auth/permission setup and XBlock handlers used across ESG views """
     authentication_classes = (
         JwtAuthentication,
         BearerAuthenticationAllowInactiveUser,
@@ -35,6 +34,78 @@ class StaffGraderBaseView(XblockHandlersMixin, RetrieveAPIView):
     )
 
     permission_classes = (IsAuthenticated,)
+
+    def get_submissions(self, request, usage_id):
+        """
+        Get a list of submissions from the ORA's 'list_staff_workflows' XBlock.json_handler
+        """
+        response = call_xblock_json_handler(request, usage_id, 'list_staff_workflows', {})
+        return json.loads(response.content)
+
+    def get_rubric_config(self, request, usage_id):
+        """
+        Get rubric data from the ORA's 'get_rubric' XBlock.json_handler
+        """
+        data = {'target_rubric_block_id': usage_id}
+        response = call_xblock_json_handler(request, usage_id, 'get_rubric', data)
+        return json.loads(response.content)
+
+    def get_submission_info(self, request, usage_id, submission_uuid):
+        """
+        Get submission content from ORA 'get_submission_info' XBlock.json_handler
+        """
+        data = {'submission_uuid': submission_uuid}
+        response = call_xblock_json_handler(request, usage_id, 'get_submission_info', data)
+        return json.loads(response.content)
+
+    def get_assessment_info(self, request, usage_id, submission_uuid):
+        """
+        Get assessment data from ORA 'get_assessment_info' XBlock.json_handler
+        """
+        data = {'submission_uuid': submission_uuid}
+        response = call_xblock_json_handler(request, usage_id, 'get_assessment_info', data)
+        return json.loads(response.content)
+
+    def submit_grade(self, request, usage_id, grade_data):
+        """
+        Submit a grade for an assessment.
+
+        Returns: {'success': True/False, 'msg': err_msg}
+        """
+        response = call_xblock_json_handler(request, usage_id, 'staff_assess', grade_data)
+        return json.loads(response.content)
+
+    def check_submission_lock(self, request, usage_id, submission_uuid):
+        """
+        Look up lock info for the given submission by calling the ORA's 'check_submission_lock' XBlock.json_handler
+        """
+        data = {'submission_uuid': submission_uuid}
+        response = call_xblock_json_handler(request, usage_id, 'check_submission_lock', data)
+        return json.loads(response.content)
+
+    def claim_submission_lock(self, request, usage_id, submission_uuid):
+        """
+        Attempt to claim a submission lock for grading.
+
+        Returns:
+        - lockStatus (string) - One of ['not-locked', 'locked', 'in-progress']
+        """
+        body = {"submission_uuid": submission_uuid}
+
+        # Return the raw response to preserve HTTP status codes for failure states
+        return call_xblock_json_handler(request, usage_id, 'claim_submission_lock', body)
+
+    def delete_submission_lock(self, request, usage_id, submission_uuid):
+        """
+        Attempt to claim a submission lock for grading.
+
+        Returns:
+        - lockStatus (string) - One of ['not-locked', 'locked', 'in-progress']
+        """
+        body = {"submission_uuid": submission_uuid}
+
+        # Return raw response to preserve HTTP status codes for failure states
+        return call_xblock_json_handler(request, usage_id, 'delete_submission_lock', body)
 
 
 class InitializeView(StaffGraderBaseView):
